Bump mdsal to 5.0.2
This bumps mdsal to 5.0.2 as part of Magnesium MRI.

+++ bgp/parser-impl/src/main/java/org/opendaylight/protocol/bgp/parser/impl/message/update/LocalPreferenceAttributeParser.java
@@ -21,6 +21,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.bgp.mess
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.bgp.message.rev180329.path.attributes.AttributesBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.bgp.message.rev180329.path.attributes.attributes.LocalPref;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.bgp.message.rev180329.path.attributes.attributes.LocalPrefBuilder;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,7 +52,7 @@ public final class LocalPreferenceAttributeParser extends AbstractAttributeParse
     public void serializeAttribute(final Attributes attribute, final ByteBuf byteAggregator) {
         final LocalPref lp = attribute.getLocalPref();
         if (lp != null) {
-            final Long pref = lp.getPref();
+            final Uint32 pref = lp.getPref();
             if (pref != null) {
                 AttributeUtil.formatAttribute(AttributeUtil.TRANSITIVE, TYPE, Unpooled.copyInt(pref.intValue()),
                     byteAggregator);
