+++ pcep/pcc-mock/src/main/java/org/opendaylight/protocol/pcep/pcc/mock/api/PCCDispatcher.java
@@ -8,20 +8,20 @@
 package org.opendaylight.protocol.pcep.pcc.mock.api;
 
 import io.netty.util.concurrent.Future;
-import java.math.BigInteger;
 import java.net.InetSocketAddress;
 import org.eclipse.jdt.annotation.NonNull;
 import org.opendaylight.protocol.concepts.KeyMapping;
 import org.opendaylight.protocol.pcep.PCEPSession;
 import org.opendaylight.protocol.pcep.PCEPSessionListenerFactory;
 import org.opendaylight.protocol.pcep.PCEPSessionNegotiatorFactory;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface PCCDispatcher {
 
     @NonNull Future<PCEPSession> createClient(@NonNull InetSocketAddress remoteAddress,
             long reconnectTime, @NonNull PCEPSessionListenerFactory listenerFactory,
             @NonNull PCEPSessionNegotiatorFactory<? extends PCEPSession> negotiatorFactory, @NonNull KeyMapping keys,
-            @NonNull InetSocketAddress localAddress, @NonNull BigInteger dbVersion);
+            @NonNull InetSocketAddress localAddress, @NonNull Uint64 dbVersion);
 
     @NonNull Future<PCEPSession> createClient(@NonNull InetSocketAddress remoteAddress,
             long reconnectTime, @NonNull PCEPSessionListenerFactory listenerFactory,
